Update Scontroller.java
I've added the methods for the options in the menu (see Client.java for a working menu). getNetstat() is currently in beta. The server can display netstat perfectly well, but I'm currently having issues with this.

NOTE: getNetstat() has a return type of String[]. I'm not currently sure if this is a good idea. The idea was supposed to be that you could take the command line response and print it elsewhere line by line. The issue is, it seems netstat doesn't work that way... I think. We need to get this working.